Connect fe with be

+++ backend/src/main/java/com/dodera/arni_fitness/service/AdminService.java
@@ -33,6 +33,7 @@ public class AdminService {
     private final SessionRepository sessionRepository;
     private final SubscriptionRepository subscriptionRepository;
     private final EntryRepository entryRepository;
+    private final ReservationRepository reservationRepository;
 
     // METODE PENTRU SUBSCRIPTII
 
@@ -59,8 +60,10 @@ public class AdminService {
     public String deleteMembership(Long id) {
         Membership membership = membershipRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException(ErrorType.MEMBERSHIP_DELETION_ERROR));
-
         try {
+            if (!membership.getPurchases().isEmpty()) {
+                throw new IllegalArgumentException("Acest tip de abonament este folosit. Nu poate fi sters.");
+            }
             Product product = Product.retrieve(membership.getStripeProductId());
             product.update(ProductUpdateParams.builder().setActive(false).build());
         } catch (StripeException e) {
@@ -372,14 +375,18 @@ public class AdminService {
     public List<SessionDetails> getSessionsDetails() {
         List<Session> sessions = sessionRepository.findAll();
 
-        return sessions.stream().map(session -> new SessionDetails(
-                session.getId(),
-                session.getName(),
-                session.getAvailableSpots(),
-                session.getSessionClassEntity().getTitle(),
-                session.getCoach().getName(),
-                session.getDatetime()
-        )).toList();
+        return sessions.stream().map(session -> {
+            List<Reservation> reservations = reservationRepository.findAllBySession(session);
+
+            return new SessionDetails(
+                    session.getId(),
+                    session.getName(),
+                    session.getAvailableSpots(),
+                    session.getSessionClassEntity().getTitle(),
+                    session.getCoach().getName(),
+                    session.getDatetime(),
+                    reservations.stream().map(Reservation::getUser).map(User::getName).toList());
+        }).toList();
     }
 
     public List<ClientDetails> getClientsDetails() {
@@ -400,6 +407,7 @@ public class AdminService {
                     user.getName(),
                     user.getEmail(),
                     user.getPhoneNumber(),
+                    user.getPin(),
                     user.getCreatedAt().toString(),
                     isActive,
                     paymentLink
@@ -443,11 +451,21 @@ public class AdminService {
         session.setDatetime(sessionRequest.date());
         session.setCoach(coach);
         session.setSessionClassEntity(classEntity);
+        session.setObservations(sessionRequest.observations());
         sessionRepository.save(session);
         return getSessionsDetails();
     }
 
     public List<SessionDetails> deleteSession(Long id) {
+        Session session = sessionRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("A aparut o eroare la stergerea antrenamentului."));
+        if (session.getDatetime().isBefore(LocalDateTime.now())) {
+            throw new IllegalArgumentException("Nu poti sterge un antrenament care a trecut deja.");
+        }
+
+        if (session.getAvailableSpots() < session.getSessionClassEntity().getAvailableSpots()) {
+            throw new IllegalArgumentException("Nu poti sterge un antrenamentul fiindca are rezervari.");
+        }
+
         sessionRepository.deleteById(id);
         return getSessionsDetails();
     }
@@ -460,21 +478,23 @@ public class AdminService {
     }
 
     public List<CoachDetails> getCoachesDetails() {
-        return getCoaches().stream().map(coach -> new CoachDetails(
-                coach.getId(),
-                coach.getName(),
-                coach.getDescription(),
-                getCoachTotalClients(coach),
-                getCoachWeeklyClients(coach),
-                coach.getCoachedClasses().stream().map(classEntity -> new CoachClassStatistics(
-                        classEntity.getId(),
-                        classEntity.getTitle(),
-                        getClassTotalClients(classEntity),
-                        getClassMonthlyClients(classEntity),
-                        getClassWeeklyClients(classEntity),
-                        getClassDailyClients(classEntity)
-                )).toList()
-        )).toList();
+        return getCoaches().stream().map(coach -> {
+            List<ClassEntity> coachedClasses = coach.getCoachedClasses();
+            return new CoachDetails(
+                    coach.getId(),
+                    coach.getName(),
+                    coach.getDescription(),
+                    getCoachTotalClients(coach),
+                    getCoachWeeklyClients(coach),
+                    coachedClasses != null ? coachedClasses.stream().map(classEntity -> new CoachClassStatistics(
+                            classEntity.getId(),
+                            classEntity.getTitle(),
+                            getClassTotalClients(classEntity),
+                            getClassMonthlyClients(classEntity),
+                            getClassWeeklyClients(classEntity),
+                            getClassDailyClients(classEntity)
+                    )).toList() : null);
+        }).toList();
     }
 
     private Integer getClassTotalClients(ClassEntity classEntity) {
@@ -497,12 +517,20 @@ public class AdminService {
     }
 
     private Integer getCoachWeeklyClients(Coach coach) {
-        return coach.getSessions().stream().filter(session -> session.getDatetime().isAfter(LocalDateTime.now().minusWeeks(1)))
+        List<Session> coachSessions = coach.getSessions();
+        if (coachSessions == null) {
+            return 0;
+        }
+        return coachSessions.stream().filter(session -> session.getDatetime().isAfter(LocalDateTime.now().minusWeeks(1)))
                 .mapToInt(session -> session.getSessionClassEntity().getAvailableSpots() - session.getAvailableSpots()).sum();
     }
 
     private Integer getCoachTotalClients(Coach coach) {
-        return coach.getSessions().stream().mapToInt(session -> session.getSessionClassEntity().getAvailableSpots() - session.getAvailableSpots()).sum();
+        List<Session> coachSessions = coach.getSessions();
+        if (coachSessions == null) {
+            return 0;
+        }
+        return coachSessions.stream().mapToInt(session -> session.getSessionClassEntity().getAvailableSpots() - session.getAvailableSpots()).sum();
     }
 
     public Item increaseQuantity(Long id) {
